requirements.txt
fastapi
uvicorn
python-dotenv
langgraph
langchain
langchain-groq
langchain-google-genai
openrouteservice
geopy
pydantic
requests